+++ lib/pdfs2pdf.rb
@@ -1,9 +1,8 @@
-# coding: utf-8
-require 'agile_utils'
-require 'code_lister'
-require_relative 'pdfs2pdf/version'
-require_relative 'pdfs2pdf/cli'
-require_relative 'pdfs2pdf/pdfs2pdf'
+require "agile_utils"
+require "code_lister"
+require_relative "pdfs2pdf/version"
+require_relative "pdfs2pdf/cli"
+require_relative "pdfs2pdf/pdfs2pdf"
 include AgileUtils::Options
 include CodeLister
 include Pdfs2Pdf