+++ test/test_helper.rb
@@ -1,8 +1,8 @@
-require 'minitest'
-require 'minitest/autorun'
-require 'minitest/pride'
-require 'minitest-spec-context'
-require 'pry'
-require 'awesome_print'
-require_relative '../lib/pdfs2pdf'
+require "minitest"
+require "minitest/autorun"
+require "minitest/pride"
+require "minitest-spec-context"
+require "pry"
+require "awesome_print"
+require_relative "../lib/pdfs2pdf"
 include Pdfs2Pdf